finish fetch, creating types game help

+++ Domain/Model/GameHelpModel/TypesGameHelpModel/HintsGameHelpModel.swift
@@ -4,10 +4,10 @@
 import Foundation
 
 public struct HintsGameHelpModel {
-    public var freeHints: Int8
-    public var adHints: Int8
+    public var freeHints: Int
+    public var adHints: Int
  
-    public init(freeHints: Int8, adHints: Int8) {
+    public init(freeHints: Int, adHints: Int) {
         self.freeHints = freeHints
         self.adHints = adHints
     }
